@@ -1,6 +1,6 @@
 {
   "name": "webpack-remove-empty-scripts",
-  "version": "0.8.1",
+  "version": "0.8.3",
   "description": "Webpack 5 plugin to remove empty scripts generated by usage only style in webpack entry.",
   "keywords": [
     "webpack",
@@ -36,21 +36,12 @@
     "access": "public",
     "registry": "https://registry.npmjs.org/"
   },
-
+  "main": "src/index.js",
   "scripts": {
     "test": "jest --detectOpenHandles --config ./test/jest.config.js",
     "test:coverage": "jest --detectOpenHandles --collectCoverage --config ./test/jest.config.js",
     "publish:public": "npm publish --access public",
     "publish:beta": "npm publish --tag beta"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "coverageDirectory": "./coverage/",
-    "collectCoverage": true
-  },
-  "main": "src/index.js",
-  "directories": {
-    "test": "test"
   },
   "files": [
     "src",
@@ -65,22 +56,22 @@
     "webpack": ">=5.32.0"
   },
   "dependencies": {
-    "ansis": "^1.3.6"
+    "ansis": "^1.4.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.5",
-    "@babel/preset-env": "^7.18.2",
-    "@types/jest": "^28.1.1",
+    "@babel/core": "^7.18.13",
+    "@babel/preset-env": "^7.18.10",
+    "@test-fixtures/lorem" : "0.0.2",
+    "@types/jest": "^28.1.8",
+    "@wordpress/scripts": "^23.4.0",
+    "@wordpress/dependency-extraction-webpack-plugin": "^4.0.0",
     "css-loader": "^6.7.1",
-    "fs": "^0.0.1-security",
-    "html-loader": "^3.1.0",
-    "jest": "^28.1.1",
-    "mini-css-extract-plugin": "^2.6.0",
-    "moment": "^2.29.3",
-    "rimraf": "^3.0.2",
-    "terser-webpack-plugin": "^5.3.3",
-    "webpack": "^5.73.0",
-    "webpack-hot-middleware": "^2.25.1",
+    "html-loader": "^4.1.0",
+    "jest": "^28.1.3",
+    "mini-css-extract-plugin": "^2.6.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "webpack": "^5.74.0",
+    "webpack-hot-middleware": "^2.25.2",
     "webpack-merge": "^5.8.0"
   }
 }
