@@ -1,87 +1,119 @@
-const fs = require('fs');
-const path = require('path');
-const rimraf = require('rimraf');
-const webpack = require('webpack');
-const { merge } = require('webpack-merge');
+import path from 'path';
+import { getCompareFileList, compareFileList, compareFileListAndContent } from './utils/helpers';
 
-import { readDirRecursiveSync } from './utils/file';
+const PATHS = {
+  base: __dirname,
+  testSource: path.join(__dirname, 'cases'),
+  // relative path in the test directory to web root dir name, same as by a web server (e.g. nginx)
+  webRoot: '/public/',
+  // relative path in the test directory to expected files for test
+  expected: '/expected/',
+  // relative path in the public directory
+  output: '/assets/',
+};
 
-//const modes = ['development'];
-const modes = ['development', 'production'];
-const cases = fs.readdirSync(path.join(__dirname, 'cases'));
+const testTimeout = 5000;
 
-// if not empty, then test only this cases
-const testOnly = [
-  //'option-extension-array',
-  //'option-extension-regexp',
-  //'option-ignore-array',
-  //'option-ignore-string',
-  // 'option-remove',
-  //'css-import',
-  //'css-entry-only',
-  //'css-entry-with-ignored-hmr',
-  //'css-entry-with-query',
-  //'multi-configuration',
-  //'multi-js-entry-css-entry',
-  //'multi-js-entry-css-entry-extensions-do-not-match',
-  //'multi-js-entry-css-entry-mjs-output-css-output',
-  //'multi-mjs-entry-css-entry',
-  //'multi-page+styles',
-  //'single-js+css-entry',
-  //'single-js+css-entry-mjs+css-output',
-  //'single-js-entry-without-ext',
-  //'single-mjs+css-entry',
-  //'single-mjs-entry',
-  //'vendor+multi-js-entry-css-entry',
-  //'webpack-concatenate-modules',
-];
-
-
-beforeAll(() => {
-  rimraf.sync(path.join(__dirname, 'output'));
-});
+beforeAll(() => {});
 
 beforeEach(() => {
-  jest.setTimeout(10000);
+  jest.setTimeout(testTimeout);
 });
 
-describe('Webpack Integration Tests', () => {
-  cases.forEach(testCase => {
-    if (testOnly.length > 0 && testOnly.indexOf(testCase) < 0) return;
+describe('production options tests', () => {
+  test('option-extension-array', (done) => {
+    compareFileListAndContent(PATHS, 'option-extension-array', done);
+  });
 
-    modes.forEach(mode => {
-      jest.setTimeout(10000);
-      test(testCase + ' [' + mode + ']', done => {
-        const testDirectory = path.join(__dirname, 'cases', testCase);
-        const outputDirectory = path.join(__dirname, 'output', testCase, mode);
-        const configFile = path.join(testDirectory, 'webpack.config.js');
+  test('option-extension-regexp', (done) => {
+    compareFileListAndContent(PATHS, 'option-extension-regexp', done);
+  });
 
-        if (!fs.existsSync(configFile)) {
-          return done('no config file for test: ' + testCase);
-        }
+  test('option-ignore-array', (done) => {
+    compareFileList(PATHS, 'option-ignore-array', done);
+  });
 
-        let baseConfig = require(configFile);
-        let configDefaults = {
-          mode: mode,
-          context: testDirectory,
-          output: {
-            path: outputDirectory,
-          },
-        };
+  test('option-ignore-string', (done) => {
+    compareFileList(PATHS, 'option-ignore-string', done);
+  });
 
-        let config = Array.isArray(baseConfig)
-          ? baseConfig.map(conf => merge(configDefaults, conf))
-          : merge(configDefaults, baseConfig);
+  test('option-remove', (done) => {
+    compareFileListAndContent(PATHS, 'option-remove', done);
+  });
+});
 
-        webpack(config, (err, stats) => {
-          if (err) return done(err);
-          if (stats.hasErrors()) return done(new Error(stats.toString()));
-          const expected = require(path.join(testDirectory, 'expected.json'));
-          const received = readDirRecursiveSync(outputDirectory, false);
-          expect(received.sort()).toEqual(expected.sort());
-          done();
-        });
-      });
-    });
+describe('production tests', () => {
+  test('css-import', (done) => {
+    compareFileListAndContent(PATHS, 'css-import', done);
+  });
+
+  test('css-entry-only', (done) => {
+    compareFileListAndContent(PATHS, 'css-entry-only', done);
+  });
+
+  test('css-entry-with-ignored-hmr', (done) => {
+    compareFileList(PATHS, 'css-entry-with-ignored-hmr', done);
+  });
+
+  test('css-entry-with-query', (done) => {
+    compareFileListAndContent(PATHS, 'css-entry-with-query', done);
+  });
+
+  test('multi-configuration', (done) => {
+    compareFileListAndContent(PATHS, 'multi-configuration', done);
+  });
+
+  test('multi-js-entry-css-entry', (done) => {
+    compareFileListAndContent(PATHS, 'multi-js-entry-css-entry', done);
+  });
+
+  test('multi-js-entry-css-entry-extensions-do-not-match', (done) => {
+    compareFileListAndContent(PATHS, 'multi-js-entry-css-entry-extensions-do-not-match', done);
+  });
+
+  test('multi-js-entry-css-entry-mjs-output-css-output', (done) => {
+    compareFileListAndContent(PATHS, 'multi-js-entry-css-entry-mjs-output-css-output', done);
+  });
+
+  test('multi-mjs-entry-css-entry', (done) => {
+    compareFileListAndContent(PATHS, 'multi-mjs-entry-css-entry', done);
+  });
+
+  test('multi-page+styles', (done) => {
+    compareFileListAndContent(PATHS, 'multi-page+styles', done);
+  });
+
+  test('single-js+css-entry', (done) => {
+    compareFileListAndContent(PATHS, 'single-js+css-entry', done);
+  });
+
+  test('single-js+css-entry-mjs+css-output', (done) => {
+    compareFileListAndContent(PATHS, 'single-js+css-entry-mjs+css-output', done);
+  });
+
+  test('single-js-entry-without-ext', (done) => {
+    compareFileListAndContent(PATHS, 'single-js-entry-without-ext', done);
+  });
+
+  test('single-mjs+css-entry', (done) => {
+    compareFileListAndContent(PATHS, 'single-mjs+css-entry', done);
+  });
+
+  test('single-mjs-entry', (done) => {
+    compareFileListAndContent(PATHS, 'single-mjs-entry', done);
+  });
+
+  test('vendor+multi-js-entry-css-entry', (done) => {
+    compareFileListAndContent(PATHS, 'vendor+multi-js-entry-css-entry', done);
+  });
+
+  test('webpack-concatenate-modules', (done) => {
+    compareFileListAndContent(PATHS, 'webpack-concatenate-modules', done);
+  });
+});
+
+describe('complex use cases with vendors', () => {
+  test('keep extracted wordpress dependencies', (done) => {
+    compareFileListAndContent(PATHS, 'wordpress-dependency', done);
   });
 });
