@@ -1,9 +1,13 @@
+const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const RemoveEmptyScriptsPlugin = require('../../../src/index.js');
 
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
 module.exports = {
-  entry: { script: './script.js', style: './style.css?global' },
+  mode: 'production',
+  output: {
+    path: path.join(__dirname, 'public/'),
+  },
+  entry: { script: './src/script.js', style: './src/style.css?global' },
   module: {
     rules: [
       {
