@@ -1,10 +1,14 @@
+const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const RemoveEmptyScriptsPlugin = require('../../../src/index.js');
 
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
 module.exports = {
+  mode: 'production',
+  output: {
+    path: path.join(__dirname, 'public/'),
+  },
   // Note: very important for test use at 0 position a css file and at 1 position a js file.
-  entry: ['./index.css', './index.js'],
+  entry: ['./src/index.css', './src/index.js'],
   module: {
     rules: [
       {
